requirements.txt
python-dotenv == 1.0.0
langchain == 0.1.0
openai == 1.7.0
unstructured == 0.11.2
tiktoken == 0.5.2
faiss-cpu == 1.7.4
streamlit == 1.29.0
langchain_openai == 0.0.2